Updated use of environment variables (closes #15)
Since issues arose at the attempt of overwriting environment variables from the host system in the container, the current solution simply declares a new set of environment variables specifying the location of the pre-trained models within the container. An additional environment variable (DEV_CONTAINER) which is only set within the container determines which of the two variables to use.